+++ StimmingSignalGenerator/Helper/RandomHelper.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 
 namespace StimmingSignalGenerator.Helper
 {
@@ -8,6 +9,7 @@ namespace StimmingSignalGenerator.Helper
       public static bool RandomBool(int percentChange) => rand.Next(0, 100) < percentChange;
       public static T GetRandomEnum<T>() where T : Enum => GetEnumValues<T>().GetRandom();
       public static T GetRandom<T>(this T[] array) => array[rand.Next(0, array.Length)];
+      public static T GetRandom<T>(this List<T> items) => items[rand.Next(0, items.Count)];
       static T[] GetEnumValues<T>() where T : Enum => (T[])Enum.GetValues(typeof(T));
    }
 }